Remove the duplicated code.
The function has two design. One is for normal use and the other one is
for test. The two part are almost the same but temp file prefix. It's
very stupid and harmful.

That's why I remove the test part code and add one extra field in test
function. The test can achieve the same goal use the normal execution
code.