Remove class SmallSeq to reduce the memory usage.
The hash function of small protein sequence is unique so the hash value
can present the orignal string. Rewrite the hash table with unsigned
integer key to reduce the meory usage. The orignal key of hash table use
8 bytes for each. The new design uses 4 bytes for each.

The other benefit for the new design is to reduce the comparing time.
The interger comparasion is much faster than string comparasion.